service impl, endpoint

+++ src/main/java/si/feri/mag/soa/model/sensors/IAmbientTemperatureSensor.java
@@ -1,5 +1,5 @@
 package si.feri.mag.soa.model.sensors;
 
 public interface IAmbientTemperatureSensor {
-    int getAmbientTemperature();
+    float getAmbientTemperature();
 }
